requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
tabulate==0.9.0
python-dateutil==2.9.0.post0
tqdm==4.66.5